Updated to link to stylesheet.
The instructions in the assignment had us create the css folder, but it didn't specify where. When going to deploy the working css the html file tried relative linking to "css/chess.css". However, that wouldn't work because the index.html file is in the a2 folder and in that folder is not where the css folder exists.

The solution here was to instruct it to look up one directory level "../css/chess.css" so it can find the css folder and the css file.